Music_tagging_task/requirements.txt
librosa==0.9.1
numpy==1.18.1
pandas==1.0.1
pydub==0.25.1
PyYAML==6.0
scikit_learn==1.1.1
scipy==1.4.1
torch==1.11.0
tqdm==4.42.1
